+++ app/routes/yara_rules.py
@@ -12,7 +12,13 @@ import datetime
 @app.route('/ThreatKB/yara_rules', methods=['GET'])
 @login_required
 def get_all_yara_rules():
-    entities = yara_rule.Yara_rule.query.all()
+    entities = yara_rule.Yara_rule.query
+
+    if not current_user.admin:
+        entities = entities.filter_by(owner_user_id=current_user.id)
+
+    entities = entities.all()
+
     return json.dumps([entity.to_dict() for entity in entities])
 
 
@@ -22,6 +28,8 @@ def get_yara_rule(id):
     entity = yara_rule.Yara_rule.query.get(id)
     if not entity:
         abort(404)
+    if not current_user.admin and entity.owner_user_id != current_user.id:
+        abort(401)
     return jsonify(entity.to_dict())
 
 
@@ -70,6 +78,8 @@ def update_yara_rule(id):
     entity = yara_rule.Yara_rule.query.get(id)
     if not entity:
         abort(404)
+    if not current_user.admin and entity.owner_user_id != current_user.id:
+        abort(401)
 
     if not do_not_bump_revision:
         db.session.add(yara_rule.Yara_rule_history(date_created=entity.date_created, revision=entity.revision,
@@ -137,6 +147,8 @@ def delete_yara_rule(id):
 
     if not entity:
         abort(404)
+    if not current_user.admin and entity.owner_user_id != current_user.id:
+        abort(401)
     db.session.delete(entity)
     db.session.commit()
